Removed by refactoring test
