+++ daft/daft/__init__.pyi
@@ -2051,13 +2051,6 @@ def set_runner_ray(
 ) -> PyDaftContext: ...
 def set_runner_native(num_threads: int | None = None) -> PyDaftContext: ...
 def get_context() -> PyDaftContext: ...
-def reset_runner() -> None:
-    """Reset/clear the current runner.
-
-    Note: This clears all in-memory state, including the partition cache.
-    """
-    ...
-
 def build_type() -> str: ...
 def version() -> str: ...
 def refresh_logger() -> None: ...
