+++ src/daft-context/src/lib.rs
@@ -110,17 +110,6 @@ impl DaftContext {
         })
     }
 
-    /// Reset/clear the current runner.
-    /// Note: This clears all in-memory state, including the partition cache.
-    pub fn reset_runner(&self) {
-        log::warn!(
-            "Resetting the runner will clear all in-memory state, including the partition cache."
-        );
-        self.with_state_mut(|state| {
-            state.runner = None;
-        });
-    }
-
     fn with_state<F, R>(&self, f: F) -> R
     where
         F: FnOnce(&ContextState) -> R,
@@ -179,12 +168,6 @@ impl DaftContext {
     }
 
     pub fn set_runner(&self, runner: Arc<Runner>) -> DaftResult<()> {
-        unimplemented!()
-    }
-
-    /// Reset/clear the current runner.
-    /// Note: This clears all in-memory state, including the partition cache.
-    pub fn reset_runner(&self) {
         unimplemented!()
     }
 
@@ -385,17 +368,6 @@ fn get_runner_config_from_env() -> DaftResult<RunnerConfig> {
 }
 
 #[cfg(feature = "python")]
-pub fn reset_runner() {
-    let ctx = get_context();
-    ctx.reset_runner();
-}
-
-#[cfg(not(feature = "python"))]
-pub fn reset_runner() {
-    unimplemented!()
-}
-
-#[cfg(feature = "python")]
 pub fn register_modules(parent: &Bound<PyModule>) -> PyResult<()> {
     parent.add_function(wrap_pyfunction!(
         python::get_runner_config_from_env,
@@ -404,7 +376,6 @@ pub fn register_modules(parent: &Bound<PyModule>) -> PyResult<()> {
     parent.add_function(wrap_pyfunction!(python::get_context, parent)?)?;
     parent.add_function(wrap_pyfunction!(python::set_runner_ray, parent)?)?;
     parent.add_function(wrap_pyfunction!(python::set_runner_native, parent)?)?;
-    parent.add_function(wrap_pyfunction!(python::reset_runner, parent)?)?;
     parent.add_class::<python::PyDaftContext>()?;
     Ok(())
 }
